Renamed job package to jobs
