+++ cmd/platform/server.go
@@ -14,7 +14,7 @@ import (
 	"github.com/mattermost/platform/api4"
 	"github.com/mattermost/platform/app"
 	"github.com/mattermost/platform/einterfaces"
-	"github.com/mattermost/platform/job"
+	"github.com/mattermost/platform/jobs"
 	"github.com/mattermost/platform/manualtesting"
 	"github.com/mattermost/platform/model"
 	"github.com/mattermost/platform/utils"
@@ -121,7 +121,7 @@ func runServer(configFileLocation string) {
 		}
 	}
 
-	jobs := job.InitJobs(app.Srv.Store).StartAll()
+	jobs := jobs.InitJobs(app.Srv.Store).StartAll()
 
 	// wait for kill signal before attempting to gracefully shutdown
 	// the running service
