+++ jobs/server.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-package job
+package jobs
 
 import (
 	l4g "github.com/alecthomas/log4go"
