+++ jobs/testjob.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-package job
+package jobs
 
 import (
 	"time"
